* trying to verify 64-bit validity of code, checking casts, especially when pointers are involved * first pass of converting c-style casts to c++ casts (static|const|reinterpret|dynamic) where possible * using misc.h to_string and from_string more often * removed unnecessary casts * using size_t more often, instead of casting to and from int/unsigned int/unsigned long * upped clientsmlcsharp project to warning level 4, fixed warnings

+++ Core/ClientSML/src/sml_DeltaList.cpp
@@ -63,7 +63,7 @@ void DeltaList::Clear(bool deleteContents)
 {
 	if (deleteContents)
 	{
-		for (int i = 0 ; i < (int)m_DeltaList.size() ; i++)
+		for (size_t i = 0 ; i < m_DeltaList.size() ; i++)
 		{
 			TagWme* pDelta = m_DeltaList[i] ;
 			delete pDelta ;
